+++ src/types/PromptbookJson/KnowledgeJson.ts
@@ -1,0 +1,8 @@
+import type { JsonFileCommon } from './JsonFileCommon';
+
+/**
+ * Knowledge contains information that is used in RAG in promptbooks
+ */
+export type KnowledgeJson = JsonFileCommon & {
+    type: 'KNOWLEDGE';
+};