+++ src/types/PromptbookJson/PromptTemplateParameterJson.ts
@@ -24,7 +24,8 @@ export type PromptTemplateParameterJson = {
 
     /**
      * Description of the parameter
-     * - It can use simple markdown formatting like **bold**, *italic*, [link](https://example.com), ... BUT not code blocks and structure
+     *
+     * Note: It can use simple markdown formatting like **bold**, *italic*, [link](https://example.com), ... BUT not code blocks and structure
      */
     readonly description?: string_markdown_text;
 };